Renamed LASReaded to openE57las, removed unused dependencies from CMake
