--- conanfile.txt
@@ -1,13 +1,0 @@
-[requires]
-boost/1.77.0
-icu/70.1
-xerces-c/3.2.3
-
-[generators]
-cmake_find_package
-
-[options]
-boost:multithreading=True
-boost:shared=False
-icu:shared=False
-xerces-c:shared=False
